requirements.txt
ccxt
pandas
numpy
scikit-learn
arch
joblib
pytorch-lightning
pyyaml
pandas-ta
streamlit
prometheus-client
python-telegram-bot
pytest